Add mutations and filters to text
